Improvements to fromdirectory ROI handling and background subtraction
ROI now defaults to NOT load an ROI file, and rather uses defaults if available, or throws an exception.

Background image assignment is now even "smarter". If it can find a molecule specific background file, it will use the last one it found for this slide, or else use the most appropriate background it can find by counting _ in filenames.