UPDATED loadstr() et al for comment parsing
Improved abstraction in loadsettings so I could create a subfunction
parseSettings to parse the comment header of any file.

Updated loaddat() to return comments as a list of lines as first element
of 3-tuple (used to be 2-tuple) including colnames and data.

Updated loadstr() to use commentparser kwarg as function to turn comment
list and return it as a data structure. If not present, output is simply
the PullData type, else its a tuple with the parser output as first
tuple element.

Created commentsToSettings() that accepts iterator with items of the
form something=value and returns a dict; piggybacks on parseSettings()
used in loadsettings().

metadata,pull = loadstr('test1.str',commentparser=commentsToSettings)