UPDATED loaddat()+dependents to pass comments as string
Comments pulled from file are passed as a string (newlines
included) but with the comment character and extra whitespace
stripped out. This isn't processed until the load() function,
which then passes that whole string to any comment parser used.
For example, comments=toSettings(), which then accepts the string
and runs splitlines() on it.

Also, reordered functions for clarity; load functions on top,
and dependancies underneath.