FIXED fix_args and validated with testing
Got nose tests running and confirmed that fix_args is working
with MS and MMS. There was a bug in useful.broadcast that was
making things complicated (left out **kwargs in one call) as well
as the fact the broadcast() mechanism required the first position
argument. I simply set a .broadcast variable on functions passing
through so fix_args would deal properly. Hack-y but works.