ADD Figure class abstraction for experiment.Pulling
Used to provide a bit nicer interface to figures
used inside of these objects. Not sure yet if it
will end up as a real class or simply for internal
use inside of experiment objects.

Specifically, Figure can at least tell you if it exists (is visible)
and also save itself to file.