Refactor collapse and aggregate in experiment.List
Also, fix saving/pickling of experiments.

aggregate() is now class method on AbstractData and allows sorting
by specific fields. This was needed because fitting to pulling
data that was simply appended would break curve fitting using masks
which are necessary for global FEC rip fits using curvefit.fitWLC_masks.

List.collapse() calls aggregate which ignores metadata since it could
all be different. It never happened that I had to check the metadata
of the aggregate because it usually only makes sense in the context
of an individual pull.