Fix sorting of filelist/grouping of experiments
fileIO.flist was returning things in wonky order. Using the filename
split on '_' as the key to sorted fixes this.

The other issue is that experiment.filelist messes up the order again.
It was using list(set()) which isn't order preserving. Added function
unique() to solve this.

Also changed how experiment.fromFile sets filename so that the
constructor has access to the filename metadata on object creation.