datatypes field access is now done by properties
Before, TrapData.ext would use __getattr__ to access extension
field in data. But with no proper __setattr__, one would mistakenly
try to do trap.ext += 100 or something but only end up setting an
attribute which wouldn't actually affect trap.data (the numpy array).

I could have implemented __setattr__ and then used an AbstractData
base method to look up and set fields... but I thought of this more
straightforward and explicit way first, which is to set each field
as a property on the object with a helper method field_property.